Updated Requested Changes
Changed maximizer to be a pawn color.
Removed unnecessary 'else' in unpack.
Changed child to child_index and the name of maximizer in alpha_beta pruning.
All other changes are in correspondence to the updated variable changes.
Tested it same as before to make sure I didn't miss anything and it still runs the same successfully.